refactor(server): extract news + earnings → news.mjs (token as param)
